services/booking_manager/requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
pydantic==2.11.4 
pydantic-settings==2.7.1
pymongo==4.12.1
python-consul==1.1.0 
httpx==0.28.1
Jinja2==3.1.6 
fastapi-mail==1.5.0 